+++ src/components/MenuHeader/index.js
@@ -1,0 +1,15 @@
+import NavBar from "./Navbar";
+import Menu from "./Menu";
+import { useState } from "react";
+
+
+const MenuHeader = () => {
+    return (
+        <>
+            <Menu />
+            <NavBar />
+        </>
+    );
+}
+
+export default MenuHeader;
